stare working

+++ app/logics/models/db/JavaInterface.java
@@ -9,7 +9,7 @@ public class JavaInterface extends JavaSourceObject {
 
 
     @IDatabaseField(columnName = "id", save = false, isID = true)
-    public int id;
+    public long id;
 
 
     @IDatabaseField(columnName = "information", fromJSON = true)
